DATA-PHYS3888_BRAIN4-simplified_version/program/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
dask==2023.4.1
distributed==2023.4.1
fonttools==4.39.3
fsspec==2023.5.0
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
keyboard==0.13.5
kiwisolver==1.4.4
llvmlite==0.40.0
locket==1.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
MouseInfo==0.1.3
msgpack==1.0.5
numba==0.57.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
partd==1.4.0
patsy==0.5.3
Pillow==9.5.0
psutil==5.9.5
PyAutoGUI==0.9.53
pygame==2.4.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc==9.1.1
pyobjc-core==9.1.1
pyobjc-framework-Accessibility==9.1.1
pyobjc-framework-Accounts==9.1.1
pyobjc-framework-AddressBook==9.1.1
pyobjc-framework-AdServices==9.1.1
pyobjc-framework-AdSupport==9.1.1
pyobjc-framework-AppleScriptKit==9.1.1
pyobjc-framework-AppleScriptObjC==9.1.1
pyobjc-framework-ApplicationServices==9.1.1
pyobjc-framework-AppTrackingTransparency==9.1.1
pyobjc-framework-AudioVideoBridging==9.1.1
pyobjc-framework-AuthenticationServices==9.1.1
pyobjc-framework-AutomaticAssessmentConfiguration==9.1.1
pyobjc-framework-Automator==9.1.1
pyobjc-framework-AVFoundation==9.1.1
pyobjc-framework-AVKit==9.1.1
pyobjc-framework-AVRouting==9.1.1
pyobjc-framework-BackgroundAssets==9.1.1
pyobjc-framework-BusinessChat==9.1.1
pyobjc-framework-CalendarStore==9.1.1
pyobjc-framework-CallKit==9.1.1
pyobjc-framework-CFNetwork==9.1.1
pyobjc-framework-ClassKit==9.1.1
pyobjc-framework-CloudKit==9.1.1
pyobjc-framework-Cocoa==9.1.1
pyobjc-framework-Collaboration==9.1.1
pyobjc-framework-ColorSync==9.1.1
pyobjc-framework-Contacts==9.1.1
pyobjc-framework-ContactsUI==9.1.1
pyobjc-framework-CoreAudio==9.1.1
pyobjc-framework-CoreAudioKit==9.1.1
pyobjc-framework-CoreBluetooth==9.1.1
pyobjc-framework-CoreData==9.1.1
pyobjc-framework-CoreHaptics==9.1.1
pyobjc-framework-CoreLocation==9.1.1
pyobjc-framework-CoreMedia==9.1.1
pyobjc-framework-CoreMediaIO==9.1.1
pyobjc-framework-CoreMIDI==9.1.1
pyobjc-framework-CoreML==9.1.1
pyobjc-framework-CoreMotion==9.1.1
pyobjc-framework-CoreServices==9.1.1
pyobjc-framework-CoreSpotlight==9.1.1
pyobjc-framework-CoreText==9.1.1
pyobjc-framework-CoreWLAN==9.1.1
pyobjc-framework-CryptoTokenKit==9.1.1
pyobjc-framework-DataDetection==9.1.1
pyobjc-framework-DeviceCheck==9.1.1
pyobjc-framework-DictionaryServices==9.1.1
pyobjc-framework-DiscRecording==9.1.1
pyobjc-framework-DiscRecordingUI==9.1.1
pyobjc-framework-DiskArbitration==9.1.1
pyobjc-framework-DVDPlayback==9.1.1
pyobjc-framework-EventKit==9.1.1
pyobjc-framework-ExceptionHandling==9.1.1
pyobjc-framework-ExecutionPolicy==9.1.1
pyobjc-framework-ExtensionKit==9.1.1
pyobjc-framework-ExternalAccessory==9.1.1
pyobjc-framework-FileProvider==9.1.1
pyobjc-framework-FileProviderUI==9.1.1
pyobjc-framework-FinderSync==9.1.1
pyobjc-framework-FSEvents==9.1.1
pyobjc-framework-GameCenter==9.1.1
pyobjc-framework-GameController==9.1.1
pyobjc-framework-GameKit==9.1.1
pyobjc-framework-GameplayKit==9.1.1
pyobjc-framework-HealthKit==9.1.1
pyobjc-framework-ImageCaptureCore==9.1.1
pyobjc-framework-IMServicePlugIn==9.1.1
pyobjc-framework-InputMethodKit==9.1.1
pyobjc-framework-InstallerPlugins==9.1.1
pyobjc-framework-InstantMessage==9.1.1
pyobjc-framework-Intents==9.1.1
pyobjc-framework-IntentsUI==9.1.1
pyobjc-framework-IOBluetooth==9.1.1
pyobjc-framework-IOBluetoothUI==9.1.1
pyobjc-framework-IOSurface==9.1.1
pyobjc-framework-iTunesLibrary==9.1.1
pyobjc-framework-KernelManagement==9.1.1
pyobjc-framework-LatentSemanticMapping==9.1.1
pyobjc-framework-LaunchServices==9.1.1
pyobjc-framework-libdispatch==9.1.1
pyobjc-framework-libxpc==9.1.1
pyobjc-framework-LinkPresentation==9.1.1
pyobjc-framework-LocalAuthentication==9.1.1
pyobjc-framework-LocalAuthenticationEmbeddedUI==9.1.1
pyobjc-framework-MailKit==9.1.1
pyobjc-framework-MapKit==9.1.1
pyobjc-framework-MediaAccessibility==9.1.1
pyobjc-framework-MediaLibrary==9.1.1
pyobjc-framework-MediaPlayer==9.1.1
pyobjc-framework-MediaToolbox==9.1.1
pyobjc-framework-Metal==9.1.1
pyobjc-framework-MetalFX==9.1.1
pyobjc-framework-MetalKit==9.1.1
pyobjc-framework-MetalPerformanceShaders==9.1.1
pyobjc-framework-MetalPerformanceShadersGraph==9.1.1
pyobjc-framework-MetricKit==9.1.1
pyobjc-framework-MLCompute==9.1.1
pyobjc-framework-ModelIO==9.1.1
pyobjc-framework-MultipeerConnectivity==9.1.1
pyobjc-framework-NaturalLanguage==9.1.1
pyobjc-framework-NetFS==9.1.1
pyobjc-framework-Network==9.1.1
pyobjc-framework-NetworkExtension==9.1.1
pyobjc-framework-NotificationCenter==9.1.1
pyobjc-framework-OpenDirectory==9.1.1
pyobjc-framework-OSAKit==9.1.1
pyobjc-framework-OSLog==9.1.1
pyobjc-framework-PassKit==9.1.1
pyobjc-framework-PencilKit==9.1.1
pyobjc-framework-PHASE==9.1.1
pyobjc-framework-Photos==9.1.1
pyobjc-framework-PhotosUI==9.1.1
pyobjc-framework-PreferencePanes==9.1.1
pyobjc-framework-PushKit==9.1.1
pyobjc-framework-Quartz==9.1.1
pyobjc-framework-QuickLookThumbnailing==9.1.1
pyobjc-framework-ReplayKit==9.1.1
pyobjc-framework-SafariServices==9.1.1
pyobjc-framework-SafetyKit==9.1.1
pyobjc-framework-SceneKit==9.1.1
pyobjc-framework-ScreenCaptureKit==9.1.1
pyobjc-framework-ScreenSaver==9.1.1
pyobjc-framework-ScreenTime==9.1.1
pyobjc-framework-ScriptingBridge==9.1.1
pyobjc-framework-SearchKit==9.1.1
pyobjc-framework-Security==9.1.1
pyobjc-framework-SecurityFoundation==9.1.1
pyobjc-framework-SecurityInterface==9.1.1
pyobjc-framework-ServiceManagement==9.1.1
pyobjc-framework-SharedWithYou==9.1.1
pyobjc-framework-SharedWithYouCore==9.1.1
pyobjc-framework-ShazamKit==9.1.1
pyobjc-framework-Social==9.1.1
pyobjc-framework-SoundAnalysis==9.1.1
pyobjc-framework-Speech==9.1.1
pyobjc-framework-SpriteKit==9.1.1
pyobjc-framework-StoreKit==9.1.1
pyobjc-framework-SyncServices==9.1.1
pyobjc-framework-SystemConfiguration==9.1.1
pyobjc-framework-SystemExtensions==9.1.1
pyobjc-framework-ThreadNetwork==9.1.1
pyobjc-framework-UniformTypeIdentifiers==9.1.1
pyobjc-framework-UserNotifications==9.1.1
pyobjc-framework-UserNotificationsUI==9.1.1
pyobjc-framework-VideoSubscriberAccount==9.1.1
pyobjc-framework-VideoToolbox==9.1.1
pyobjc-framework-Virtualization==9.1.1
pyobjc-framework-Vision==9.1.1
pyobjc-framework-WebKit==9.1.1
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pyserial==3.5
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0
requests==2.30.0
rubicon-objc==0.4.6
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sortedcontainers==2.4.0
statsmodels==0.14.0
stumpy==1.11.1
tblib==1.7.0
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
tsfresh==0.20.0
tzdata==2023.3
urllib3==2.0.2
zict==3.0.0
zipp==3.15.0
